Software/sandbox/robot-arm-3d-sim/requirements.txt
numpy>=1.24
PyYAML>=6.0
matplotlib>=3.8
